+++ packages/ui-core/src/components/Button/Button.test.tsx
@@ -1,14 +1,25 @@
 import * as React from 'react';
 import { shallow, mount } from 'enzyme';
-import Button from './Button';
+import Button, { IButtonProps } from './Button';
+
+const props: IButtonProps = {
+    type: 'reset',
+    href: 'any',
+    target: '_blank',
+};
 
 describe('<Button />', () => {
     beforeEach(() => {
         jest.resetModules();
     });
 
-    it('it renders Button', () => {
+    it('renders Button with default props', () => {
         const wrapper = shallow(<Button />);
+        expect(wrapper).toMatchSnapshot();
+    });
+
+    it('renders Button with props', () => {
+        const wrapper = shallow(<Button {...props} />);
         expect(wrapper).toMatchSnapshot();
     });
 
@@ -35,16 +46,44 @@ describe('<Button />', () => {
         expect(wrapper).toMatchSnapshot();
     });
 
-    it('it renders tag a', () => {
-        const wrapper = mount(<Button href="http://moscow.megafon.ru/" />);
-        expect(wrapper).toMatchSnapshot();
+    describe('with disabled state', () => {
+        it('renders tag button', () => {
+            const wrapper = mount(<Button customView={undefined} disabled href={undefined} />);
+            expect(wrapper).toMatchSnapshot();
+        });
+
+        it('renders tag a', () => {
+            const wrapper = mount(<Button customView={undefined} disabled href="test" />);
+            expect(wrapper).toMatchSnapshot();
+        });
+
+        it('renders element with customView', () => {
+            const wrapper = mount(<Button customView="two-lines" disabled />);
+            expect(wrapper).toMatchSnapshot();
+        });
     });
 
-    it('it calls click handler', () => {
-        const onClick = jest.fn();
-        const wrapper = mount(<Button href="http://moscow.megafon.ru/" onClick={onClick} />);
-        wrapper.simulate('click');
-        expect(onClick).toHaveBeenCalledTimes(1);
+    describe('handleClick', () => {
+        it('calls when disabled is false', () => {
+            const onClick = jest.fn();
+            const preventDefault = jest.fn();
+            const wrapper = shallow(<Button onClick={onClick} />);
+
+            wrapper.simulate('click', { preventDefault });
+
+            expect(onClick).toBeCalledWith({ preventDefault });
+            expect(preventDefault).not.toBeCalled();
+        });
+
+        it('calls when disabled is true', () => {
+            const onClick = jest.fn();
+            const preventDefault = jest.fn();
+            const wrapper = shallow(<Button disabled onClick={onClick} />);
+
+            wrapper.simulate('click', { preventDefault });
+            expect(onClick).not.toBeCalled();
+            expect(preventDefault).toBeCalled();
+        });
     });
 
     it('it passes props to mods', () => {