+++ packages/ui-core/CHANGELOG.md
@@ -4,6 +4,12 @@
 - **GridColumn**
     - add box-sizing: border-box in less
     - fix gutters width and height for mobile devices
+- **Button**
+    - fixed disabled state for tag a (block onClick calls, cancel following link, change color)
+    - fixed display target attribute for tag button
+    - fixed display type attribute for tag a
+    - added tests
+    - added usage example for disabled, href, target
 
 ## Release 1.0.20
 - **Icons**